+++ pybamm/expression_tree/symbol.py
@@ -232,6 +232,26 @@ class Symbol:
             ):
                 self.test_shape()
 
+    @classmethod
+    def _from_json(cls, snippet: dict):
+        """
+        Reconstructs a Symbol instance during deserialisation of a JSON file.
+
+        Parameters
+        ----------
+        snippet: dict
+            Contains the information needed to reconstruct a specific instance.
+            At minimum, should contain "name", "children" and "domains".
+        """
+
+        instance = cls.__new__(cls)
+
+        instance.__init__(
+            snippet["name"], children=snippet["children"], domains=snippet["domains"]
+        )
+
+        return instance
+
     @property
     def children(self):
         """
@@ -986,3 +1006,16 @@ class Symbol:
     def to_equation(self):
         sympy = have_optional_dependency("sympy")
         return sympy.Symbol(str(self.name))
+
+    def to_json(self):
+        """
+        Method to serialise a Symbol object into JSON.
+        """
+
+        json_dict = {
+            "name": self.name,
+            "id": self.id,
+            "domains": self.domains,
+        }
+
+        return json_dict