+++ tests/unit/test_expression_tree/test_parameter.py
@@ -31,6 +31,15 @@ class TestParameter(TestCase):
         # Test name
         self.assertEqual(func1.to_equation(), sympy.Symbol("test_name"))
 
+    def test_to_json_error(self):
+        func = pybamm.Parameter("test_string")
+
+        with self.assertRaises(NotImplementedError):
+            func.to_json()
+
+        with self.assertRaises(NotImplementedError):
+            pybamm.Parameter._from_json({})
+
 
 class TestFunctionParameter(TestCase):
     def test_function_parameter_init(self):
@@ -109,6 +118,15 @@ class TestFunctionParameter(TestCase):
         # Test name
         func1.print_name = None
         self.assertEqual(func1.to_equation(), sympy.Symbol("func"))
+
+    def test_to_json_error(self):
+        func = pybamm.FunctionParameter("test", {"x": pybamm.Scalar(1)})
+
+        with self.assertRaises(NotImplementedError):
+            func.to_json()
+
+        with self.assertRaises(NotImplementedError):
+            pybamm.FunctionParameter._from_json({})
 
 
 if __name__ == "__main__":